Added type serializer + Fixed circular import
